renamed placetubespage to place span equipments
