+++ src/pages/PlaceSpanEquipmentPage/PlaceSpanEquipmentGql.ts
