+++ src/pages/PlaceSpanEquipmentPage/index.ts
@@ -1,0 +1,4 @@
+import PlaceSpanEquipmentPage from "./PlaceSpanEquipmentPage";
+import "./PlaceSpanEquipmentPage.scss";
+
+export default PlaceSpanEquipmentPage;
