--- src/pages/PlaceTubesPage/index.ts
@@ -1,4 +1,0 @@
-import PlaceTubesPage from "./PlaceTubesPage";
-import "./PlaceTubesPage.scss";
-
-export default PlaceTubesPage;
