+++ src/translation/translation.json
@@ -4,6 +4,7 @@
       "Home": "Home",
       "Welcome to OpenFTTH": "Welcome to OpenFTTH",
       "Place span equipment": "Place span equipment",
+      "Place span equipments": "Place span equipments",
       "Manufacturer": "Manufacturer",
       "Specification": "Specification",
       "Pick color marking": "Pick color marking",
@@ -35,6 +36,7 @@
       "Home": "Hjem",
       "Welcome to OpenFTTH": "Velkommen til OpenFTTH",
       "Place span equipment": "Placer ledning",
+      "Place span equipments": "Placer ledninger",
       "Manufacturer": "Fabrikant",
       "Specification": "Specifikation",
       "Pick color marking": "Vælg farvekode",
